pyproject.toml
[tool.poetry]
authors = ["long2ice <long2ice@gmail.com>"]
description = "A light alternative for Sentry"
name = "guard"
version = "0.1.0"

[tool.poetry.dependencies]
python = "^3.7"
fastapi = "*"
tortoise-orm = "*"
uvicorn = "*"
dynaconf = "*"
asyncmy = "*"
asyncpg = "*"
aioredis = "*"

[tool.poetry.dev-dependencies]
# test
pytest = "*"
pytest-asyncio = "*"
pytest-mock = "*"
pytest-sugar = "*"
pytest-xdist = "*"
# lint
black = "*"
flake8 = "*"
isort = "*"
mypy = "*"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]
